+++ R/predict.fastNaiveBayes.multinomial.R
@@ -89,6 +89,7 @@ predict.fastNaiveBayes.multinomial <- function(object, newdata, type = c("class"
   }
 
   denom <- rowSums(probs)
+  denom[denom==0] <- 1
   probs <- probs / denom
 
   if (type == "class") {